requirements.txt
tensorflow>=2.9.0
opencv-python
numpy
Pillow
matplotlib
streamlit